chore: refactoring the nostr code
